feat(datasets): add multi_en ASR recipe and audio datasets package
Port result/data/build_train_mix.py into a first-class
make_multi_en_asr_dataset builder (six-corpus English mix; torchcodec-free
soundfile decode) and reorganize audio/ASR dataset code into a new
components/datasets/audio/ package.

- New components/datasets/label_utils.py holds the shared label/marker chain
  (incl. default_stop_tokens), so audio never imports vlm. vlm/collate_fns.py
  re-imports those names (string-path monkeypatches keep working) and imports
  the omni-ASR collates from audio to populate COLLATE_FNS (one-directional).
- audio/datasets.py: moved make_hf_audio_asr_dataset, make_cv17_dataset,
  _decode_audio_cell_to_mono_float32, _build_asr_conversation; extracted a
  reusable _attach_asr_transform.
- audio/multi_en.py: make_multi_en_asr_dataset (+ build_multi_en_source_mix),
  prototype normalize_text, six-source spec, empty-after-normalization filter.
- audio/collate_fns.py: moved qwen2_5/qwen3 omni ASR collates + helpers.
- examples/audio_finetune/qwen2_5_omni_asr/multi_en_sft_3b.yaml +
  train_multi_en_qwen2.5_omni_3b.sh (single-node 8-GPU launcher).
- Updated all moved-symbol references (YAMLs, docs + Fern mirrors); no shims.
- Relocated audio/ASR + label-builder tests; new test_multi_en.py.
