+++ examples/audio_finetune/qwen2_5_omni_asr/ami_sft_7b.yaml
@@ -74,7 +74,7 @@ loss_fn:
 #   user(  text=user_prompt + audio=<waveform>  )
 #   assistant(  text=<transcript>  )
 dataset:
-  _target_: nemo_automodel.components.datasets.vlm.datasets.make_hf_audio_asr_dataset
+  _target_: nemo_automodel.components.datasets.audio.datasets.make_hf_audio_asr_dataset
   path_or_dataset: edinburghcstr/ami
   name: ihm
   split: train
@@ -94,7 +94,7 @@ dataloader:
   _target_: torchdata.stateful_dataloader.StatefulDataLoader
   num_workers: 2
   collate_fn:
-    _target_: nemo_automodel.components.datasets.vlm.collate_fns.qwen2_5_omni_asr_collate_fn
+    _target_: nemo_automodel.components.datasets.audio.collate_fns.qwen2_5_omni_asr_collate_fn
 
 optimizer:
   _target_: torch.optim.AdamW
