+++ examples/audio_finetune/qwen3_omni_asr/ami_sft.yaml
@@ -101,7 +101,7 @@ loss_fn:
 #   assistant(  text=<transcript>  )
 # No system turn (system_prompt: null).
 dataset:
-  _target_: nemo_automodel.components.datasets.vlm.datasets.make_hf_audio_asr_dataset
+  _target_: nemo_automodel.components.datasets.audio.datasets.make_hf_audio_asr_dataset
   path_or_dataset: edinburghcstr/ami
   name: ihm
   split: train
@@ -118,7 +118,7 @@ dataloader:
   _target_: torchdata.stateful_dataloader.StatefulDataLoader
   num_workers: 2
   collate_fn:
-    _target_: nemo_automodel.components.datasets.vlm.collate_fns.qwen3_omni_asr_collate_fn
+    _target_: nemo_automodel.components.datasets.audio.collate_fns.qwen3_omni_asr_collate_fn
 
 optimizer:
   _target_: torch.optim.AdamW
