+++ nemo_automodel/components/datasets/audio/__init__.py
@@ -1,0 +1,31 @@
+# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from nemo_automodel.components.datasets.audio.collate_fns import (
+    qwen2_5_omni_asr_collate_fn,
+    qwen3_omni_asr_collate_fn,
+)
+from nemo_automodel.components.datasets.audio.datasets import (
+    make_cv17_dataset,
+    make_hf_audio_asr_dataset,
+)
+from nemo_automodel.components.datasets.audio.multi_en import make_multi_en_asr_dataset
+
+__all__ = [
+    "make_hf_audio_asr_dataset",
+    "make_cv17_dataset",
+    "make_multi_en_asr_dataset",
+    "qwen2_5_omni_asr_collate_fn",
+    "qwen3_omni_asr_collate_fn",
+]
